Refactor image compression function and import statement in UploadImages component
